# openapi.yaml
openapi: 3.0.3
info:
  title: Anime Scrapper API
  version: '2.0.0'
  description: |
    API publica read-only para consultar animes y episodios precalculados en Supabase.
    Esta API no ejecuta scraping en runtime.
servers:
  - url: https://{host}
    description: Entorno dinamico (dev/prod)
    variables:
      host:
        default: localhost:3000
        description: Host de desarrollo
  - url: https://anime-scrapper.vercel.app
    description: Produccion (ejemplo)
paths:
  /api/animes:
    get:
      summary: Listado de animes del feed directory
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
          required: false
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 24 }
          required: false
      responses:
        '200':
          description: Lista de animes
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Anime' }
        '404':
          description: No se encontraron animes
          content:
            text/plain: { schema: { type: string } }
        '500':
          description: Error servidor
          content:
            text/plain: { schema: { type: string } }
...
  /api/animes/{animeId}:
    get:
      summary: Obtener detalle de anime
      parameters:
        - name: animeId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Detalle de anime
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Anime' }
        '404':
          description: Anime no encontrado
          content:
            text/plain: { schema: { type: string } }
        '500':
          description: Error servidor
          content:
            text/plain: { schema: { type: string } }
...
components:
  schemas:
    Anime:
      type: object
      properties:
        animeId: { type: string, example: 'solo-leveling' }
        title: { type: string, example: 'Solo Leveling' }
        images:
          type: object
          nullable: true
          properties:
            coverImage: { type: string, format: uri, nullable: true }
            carouselImages:
              type: array
              items:
                $ref: '#/components/schemas/CarouselImage'
        type: { type: string, nullable: true, example: 'TV' }
        rank: { type: number, nullable: true, example: 1 }
        otherTitles:
          type: array
          items: { type: string }
          nullable: true
        description: { type: string, nullable: true }
        originalLink: { type: string, nullable: true, format: uri }
        status: { type: string, nullable: true, example: 'En emision' }
        genres:
          type: array
          items: { type: string }
          nullable: true
        relatedAnimes:
          type: array
          items: { $ref: '#/components/schemas/RelatedAnime' }
          nullable: true
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [animeId, title, created_at, updated_at]
    CarouselImage:
      type: object
      properties:
        link: { type: string, nullable: true, format: uri }
        position: { type: string, example: '50% 50%' }
        width: { type: integer, example: 1536 }
        height: { type: integer, example: 768 }
      required: [position, width, height]
    RelatedAnime:
      type: object
      properties:
        animeId: { type: string }
        title: { type: string }
        relation: { type: string }
      required: [animeId, title, relation]
...